+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fixed a bug where you couldn't view the parent note of a reply when it was displayed at the top of the Thread view.
 - Updated support email to support@nos.social
 - Improved recognition of mentions inside a post
 - Fixed a bug where expired messages could be redownloaded from relays that don't delete them.